Make it such that banner and navbar don't show when survey is open and in logged out state
