+++ src/components/Navbar/index.tsx
@@ -15,7 +15,17 @@ import { useState, useEffect, useRef } from "react";
 import { useSubscription } from "../../hooks/useSubscription.ts";
 import GetRoleSurvey from "../Survey/getRole";
 
-export default function Navbar() {
+export default function Navbar({
+  surveyVariables,
+}: {
+  surveyVariables: {
+    hasDoneSurvey: boolean;
+    surveyClosed: boolean;
+    onSurveySubmit: (role: string, isNoResponse?: boolean) => void;
+    onClose: () => void;
+    setShowSurvey: (show: boolean) => void;
+  };
+}) {
   const { signOut } = useClerk();
   const cartItems = useReactiveVar(cartItemsVar);
   const [location, setLocation] = useLocation();
@@ -55,9 +65,9 @@ export default function Navbar() {
   return (
     <div className="sticky top-0 z-10 p-4">
       <div className="navbar bg-primary rounded-box">
-        <div className="navbar-start flex items-center">
+        <div className="flex items-center navbar-start">
           <Link href={`/practice${queryString ? `?${queryString}` : ""}`}>
-            <div className="text-xl mr-2">Teebloc</div>
+            <div className="mr-2 text-xl">Teebloc</div>
           </Link>
           <div className="dropdown">
             <label tabIndex={0} className="m-1 btn btn-ghost">
@@ -129,6 +139,7 @@ export default function Navbar() {
                 location={location}
                 cartItems={cartItems}
                 onSignOut={onSignOut}
+                surveyVariables={surveyVariables}
               />
             </div>
           )}
@@ -139,6 +150,7 @@ export default function Navbar() {
             location={location}
             cartItems={cartItems}
             onSignOut={onSignOut}
+            surveyVariables={surveyVariables}
           />
         </div>
       </div>
@@ -151,10 +163,18 @@ function NavItems({
   location,
   cartItems,
   onSignOut,
+  surveyVariables,
 }: {
   location: string;
   cartItems: any[]; // Replace 'any' with the correct type if known
   onSignOut: () => void;
+  surveyVariables: {
+    hasDoneSurvey: boolean;
+    surveyClosed: boolean;
+    onSurveySubmit: (role: string, isNoResponse?: boolean) => void;
+    onClose: () => void;
+    setShowSurvey: (show: boolean) => void;
+  };
 }) {
   const { getToken } = useAuth();
   const [writingMatch] = useRoute("/writing");
@@ -179,45 +199,7 @@ function NavItems({
     });
   };
 
-  const SURVEY_ID = import.meta.env.VITE_USER_ROLE_SURVEY_ID;
-  const SURVEY_QUESTION_ID = import.meta.env.VITE_USER_ROLE_SURVEY_QUESTION_ID;
-
-  const [showSurvey, setShowSurvey] = useState(false);
-  const [showThanks, setShowThanks] = useState(false);
-  const [hasDoneSurvey, setHasDoneSurvey] = useState(
-    localStorage.getItem(`hasInteractedWithSurvey_${SURVEY_ID}`) === "true"
-  );
-  const [surveyClosed, setSurveyClosed] = useState(false);
-
-  const onClose = () => {
-    setShowSurvey(false);
-    setShowThanks(false);
-    setSurveyClosed(true);
-  };
-
-  const onSurveySubmit = (role: string, isNoResponse: boolean = false) => {
-    setHasDoneSurvey(true);
-    setShowSurvey(false);
-    if (!isNoResponse) setShowThanks(true);
-
-    const responseKey = `$survey_response_${SURVEY_QUESTION_ID}`;
-
-    posthog.capture("survey sent", {
-      $survey_id: SURVEY_ID,
-      [responseKey]: isNoResponse ? "No response" : role,
-    });
-
-    localStorage.setItem(`hasInteractedWithSurvey_${SURVEY_ID}`, "true");
-  };
-
-  useEffect(() => {
-    if (surveyClosed) {
-      const signUpBtn = document.getElementById("sign-up-button");
-      if (signUpBtn) {
-        signUpBtn.click();
-      }
-    }
-  }, [surveyClosed]);
+  const { hasDoneSurvey, surveyClosed, onSurveySubmit, onClose, setShowSurvey } = surveyVariables;
 
   return (
     <>
@@ -240,14 +222,6 @@ function NavItems({
           </button>
         )}
       </SignedOut>
-      {(showSurvey || showThanks) && (
-        <GetRoleSurvey
-          submitHandler={onSurveySubmit}
-          onClose={onClose}
-          onCancel={onClose}
-          hasDoneSurvey={hasDoneSurvey}
-        />
-      )}
       <SignedIn>
         {!hasDoneSurvey && (
           <GetRoleSurvey
